refactor: remove legacy MVC
